eliminate RevsisionManager
